# servers-integration-api.openapi+json.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Servers API",
        "version": "v1"
    },
    "paths": {
        "/maps/{gameServerId}/host/loaded": {
            "get": {
                "tags": [
                    "Maps"
                ],
                "parameters": [
                    {
                        "name": "gameServerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/maps/{gameServerId}/host/{mapName}": {
            "post": {
                "tags": [
                    "Maps"
                ],
                "parameters": [
                    {
                        "name": "gameServerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "mapName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/query/{gameServerId}/status": {
            "get": {
                "tags": [
                    "Query"
                ],
                "parameters": [
                    {
                        "name": "gameServerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/rcon/{gameServerId}/status": {
            "get": {
                "tags": [
                    "Rcon"
                ],
                "parameters": [
                    {
                        "name": "gameServerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/rcon/{gameServerId}/maps": {
            "get": {
                "tags": [
                    "Rcon"
                ],
                "parameters": [
                    {
                        "name": "gameServerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/": {
            "get": {
                "tags": [
                    "servers-integration-webapi"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}